wip(podgangmap): rewrite PodGangMap component (podgangmap-new) with tests
Intermediate checkpoint. Adds the rewritten PodGangMap component under
podgangmap-new (bootstrap, steady-state, reconstruction, coherent-update
sub-step planning + materialization) with full unit tests, test/utils
builders, and swaps the registry to it. Orchestrator, PodGang materializer,
and PCLQ/PCSG consumer changes are still pending, so the module does not
fully compile yet (old podgangmap package and podgang_test.go carry
pre-existing ListExistingPodGangs signature drift).

+++ operator/internal/controller/podcliquescalinggroup/components/podclique/pcsgreplicaflow.go
@@ -200,7 +200,7 @@ func generateScaledPodGangNames(count int, pcsName string, pcsReplicaIndex int, 
 	names := make([]string, 0, count)
 	for i := range count {
 		suffix := strconv.FormatInt(base+int64(i), 10)
-		names = append(names, apicommon.GeneratePodGangName(pcsName, int32(pcsReplicaIndex), suffix))
+		names = append(names, apicommon.GeneratePodGangName(pcsName, pcsReplicaIndex, suffix))
 	}
 	return names
 }
